refactor: 兼容专用的内存管理库

+++ Sevm.Engine/ScriptRegisters.cs
@@ -7,7 +7,7 @@ namespace Sevm.Engine {
     /// <summary>
     /// 寄存器集合
     /// </summary>
-    public class ScriptRegisters : List<int> {
+    public class ScriptRegisters : List<long> {
 
         // 自动扩展集合
         private void AutoCreate(int index) {
@@ -23,7 +23,7 @@ namespace Sevm.Engine {
         /// </summary>
         /// <param name="index"></param>
         /// <returns></returns>
-        public new int this[int index] {
+        public new long this[int index] {
             get {
                 this.AutoCreate(index);
                 return base[index];
